+++ R/internals.R
@@ -86,17 +86,17 @@ vis_extract_value_ <- function(x){
 vis_create_ <- function(x){
 
   ggplot2::ggplot(data = x,
-                  ggplot2::aes_string(x = "variable",
-                                      y = "rows",
-                                      # text assists with plotly mouseover
-                                      text = "value")) +
-    ggplot2::geom_raster(ggplot2::aes_string(fill = "valueType")) +
-    ggplot2::theme_minimal() +
-    ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45,
-                                                       vjust = 1,
-                                                       hjust = 1)) +
-    ggplot2::labs(x = "",
-                  y = "Observations") +
+                  ggplot2::aes(x = variable,
+                               y = rows,
+                                    # text assists with plotly mouseover
+                                    text = value)) +
+  ggplot2::geom_raster(ggplot2::aes(fill = valueType)) +
+  ggplot2::theme_minimal() +
+  ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45,
+                                                     vjust = 1,
+                                                     hjust = 1)) +
+  ggplot2::labs(x = "",
+                y = "Observations") +
     # flip the axes
     ggplot2::scale_y_reverse() +
     ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.5)) +
@@ -358,12 +358,17 @@ test_if_all_binary <- function(data){
   }
 }
 
+monotonic <- function(x) all(diff(x) == 0)
+
 #' Scale a vector between 0 and one.
 #'
 #' @param x numeric vector
 #'
 #' @return numeric vector between 0 and 1
 scale_01 <- function(x) {
+  if (monotonic(x)){
+    return((x/x))
+  }
   (x - min(x, na.rm = TRUE)) / diff(range(x, na.rm = TRUE))
 }